Refactor thread code into threaded.rs
